pyproject.toml
[project]
name = "nl2sql"
version = "0.1.0"
description = "A modular pipeline that converts natural language questions into SQL queries using LLMs via Z.AI API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "openai>=1.0.0",
    "pydantic>=2.0",
    "pydantic-core>=2.41.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlparse>=0.5.5",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
